Move navigation color tests to own file
